fix(storefront): Make logout always clear the session and keep customer keys consistent
Remaining review points from PR #808:
- `emptySession` shared mutable constant replaced by a factory:
  `useStorage` may keep the initial value as the live reactive state, so
  the `logout()` reset was a self-assignment no-op (token surviving and
  UI still logged) and later writes polluted the "empty" constant with
  the previous customer's data;
- `fetchCustomer` now keeps the `display_name`/`main_email` always-set
  invariant that consumers assume, as the boot purge already did;
- Censored session helpers extracted to `state/customer-session/` per
  directory convention (as `state/shopping-cart/`).
